cleanup

+++ covidash/data/download.py
@@ -4,12 +4,18 @@ from pathlib import Path
 
 DATA_DIR = '../../data/raw'
 
+def get_rki_data(url="", file_name=""):
+    file_name = os.path.join(DATA_DIR, file_name)
+    if not Path(file_name).is_file():
+        urllib.request.urlretrieve(url, file_name)
+
+def download_osm_hospital_locations():
+    get_rki_data(url="https://opendata.arcgis.com/datasets/348b643c8b234cdc8b1b345210975b87_0.geojson",
+                 file_name="osm_hospital_locations_germany.json")
 
 def download_rki_covid_19():
-    url = 'https://opendata.arcgis.com/datasets/dd4580c810204019a7b8eb3e0b329dd6_0.csv'
-    file_name = os.path.join(DATA_DIR, 'rki_covid_19.csv')
-    urllib.request.urlretrieve(url, file_name)
-
+    get_rki_data(url = 'https://opendata.arcgis.com/datasets/dd4580c810204019a7b8eb3e0b329dd6_0.csv',
+                 file_name='rki_covid_19.csv')
 
 def download_rki_corona_landkreise():
     url = 'https://opendata.arcgis.com/datasets/917fc37a709542548cc3be077a786c17_0.csv'
@@ -24,15 +30,10 @@ def download_rki_corona_laendergrenzen():
     urllib.request.urlretrieve(url, file_name)
 
 
-def get_rki_data(url="", file_name=""):
-    file_name = os.path.join(DATA_DIR, file_name)
-    if not Path(file_name).is_file():
-        urllib.request.urlretrieve(url, file_name)
 
 
-def download_osm_hospital_locations():
-    get_rki_data(url="https://opendata.arcgis.com/datasets/348b643c8b234cdc8b1b345210975b87_0.geojson",
-                 file_name="osm_hospital_locations_germany.json")
+
+
 
 
 if __name__ == '__main__':
